# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "hotspot_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotspot Id"
          },
          "locale": {
            "const": "zh-CN",
            "default": "zh-CN",
            "title": "Locale",
            "type": "string"
          },
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question",
          "conversation_id"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "insufficient_evidence": {
            "title": "Insufficient Evidence",
            "type": "boolean"
          },
          "local_action_suggestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Action Suggestion"
          },
          "provider": {
            "enum": [
              "mock",
              "openai"
            ],
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "citations",
          "insufficient_evidence",
          "provider"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "Citation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "type": {
            "enum": [
              "knowledge",
              "article"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "FeedbackEvent": {
        "additionalProperties": false,
        "properties": {
          "client_event_id": {
            "format": "uuid",
            "title": "Client Event Id",
            "type": "string"
          },
          "event_type": {
            "enum": [
              "click",
              "read",
              "collect",
              "hide"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "hotspot_id": {
            "format": "uuid",
            "title": "Hotspot Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          }
        },
        "required": [
          "client_event_id",
          "hotspot_id",
          "event_type",
          "occurred_at"
        ],
        "title": "FeedbackEvent",
        "type": "object"
      },
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/FeedbackEvent"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "FeedbackResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "公考晨报 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/ai/chat": {
      "post": {
        "operationId": "chat_v1_ai_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-client-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Chat"
      }
    },
...
    "/v1/feedback/events": {
      "post": {
        "operationId": "feedback_v1_feedback_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-client-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Feedback"
      }
    },
...
  }
}